Moved configurable features to a separate folder
